+++ CHANGELOG.md
@@ -10,6 +10,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - `position` action to product summary context
 
+## [0.11.0] - 2023-12-20
+
+### Added
+
+- Added the argument `sponsoredBadgeOptions` to `ProductSummaryProvider`.
+
 ## [0.10.0] - 2023-09-27
 
 ## [0.9.0] - 2021-09-06